Moved tests to regression
