+++ src/EPPlus.Fonts.OpenType.Tests/FontMeasurerPerformanceTest.cs
@@ -134,49 +134,5 @@ namespace EPPlus.Fonts.OpenType.Tests
             //Assert.AreEqual(0, differingStrings.Count());
             //Assert.AreEqual(UnOptimizedOriginalWrappingString, currStr);
         }
-        [TestMethod]
-        public void SettingBoldItalicAgainShouldNotTimeout()
-        {
-            MeasurementFont boldItalic = new MeasurementFont()
-            {
-                FontFamily = "Aptos Narrow",
-                Size = 11f,
-                Style = MeasurementFontStyles.Bold | MeasurementFontStyles.Italic
-            };
-
-            var ttTextMeasurer = new FontMeasurerTrueType();
-
-            Stopwatch timer = new Stopwatch();
-            timer.Start();
-            ttTextMeasurer.SetFont(boldItalic);
-            timer.Stop();
-            var firstTime = timer.ElapsedMilliseconds;
-
-            timer.Restart();
-            ttTextMeasurer.SetFont(boldItalic);
-            timer.Stop();
-
-            //Doing the same operation again should not be a whole second longer
-            Assert.IsTrue((firstTime + 1000) > timer.ElapsedMilliseconds);
-            //At time of writing OpenTypeFontCache.GetFromCache is 2s therefore it should take less
-            Assert.IsTrue(timer.ElapsedMilliseconds < 2000);
-        }
-
-        [TestMethod]
-        public void GetFromCacheBoldItalicShouldWork()
-        {
-            MeasurementFont boldItalic = new MeasurementFont()
-            {
-                FontFamily = "Aptos Narrow",
-                Size = 11f,
-                Style = MeasurementFontStyles.Bold | MeasurementFontStyles.Italic
-            };
-
-            var ttTextMeasurer = new FontMeasurerTrueType();
-            ttTextMeasurer.SetFont(boldItalic);
-
-            var cachedFont = OpenTypeFontCache.GetFromCache("Aptos Narrow", FontSubFamily.BoldItalic);
-            Assert.IsNotNull(cachedFont);
-        }
     }
 }
